Add new service-worker
